server: clean up the intermediate steps in createContainer
