+++ server/container_create_generic.go
@@ -1,0 +1,10 @@
+// +build windows darwin
+
+package server
+
+import "github.com/kubernetes-incubator/cri-o/oci"
+
+// createContainerPlatform performs platform dependent intermediate steps before calling the container's oci.Runtime().CreateContainer()
+func (s *Server) createContainerPlatform(container *oci.Container, infraContainer *oci.Container, cgroupParent string) error {
+	return s.Runtime().CreateContainer(container, cgroupParent)
+}
