+++ server/sandbox_run_linux.go
@@ -589,7 +589,7 @@ func (s *Server) runPodSandbox(ctx context.Context, req *pb.RunPodSandboxRequest
 		}
 	}()
 
-	if err = s.createContainer(container, nil, sb.CgroupParent()); err != nil {
+	if err = s.createContainerPlatform(container, nil, sb.CgroupParent()); err != nil {
 		return nil, err
 	}
